Merge the prototype bootstrap UI into the laravel-generated code
The UI represented in this commit is a modified version of the Regna
template created by boostrapmade.com:

https://bootstrapmade.com/regna-bootstrap-onepage-template/

A hit on index.html will display this code. One (and only one) link
inside redirects to laravel.php as a prototype bridge between the
static and the dynamic.

At this very early stage, I have no idea whatsoever how this will
work. This is, in every sense, a mockup.